some rough permissions handling

+++ main/views.py
@@ -10,12 +10,14 @@ from datetime import datetime
 @login_required
 @render_to("main/index.html")
 def node(request,node_id):
-    due = Node.objects.filter(target=datetime.now())
-    overdue = Node.objects.filter(target__lt=datetime.now()).order_by("target")
-    upcoming = Node.objects.filter(target__gt=datetime.now()).order_by("target")[:10]
+    due = Node.objects.filter(user=request.user,target=datetime.now())
+    overdue = Node.objects.filter(user=request.user,target__lt=datetime.now()).order_by("target")
+    upcoming = Node.objects.filter(user=request.user,target__gt=datetime.now()).order_by("target")[:10]
     d = dict(due=due,overdue=overdue,upcoming=upcoming)
     if node_id:
         n = get_object_or_404(Node,id=node_id)
+        if n.user.id != request.user.id:
+            return HttpResponse("you are not the owner of this node")
         d['node'] = n
     return d
 
@@ -23,13 +25,16 @@ def node(request,node_id):
 @render_to("main/search.html")
 def search(request):
     q = request.GET.get("q","")
-    results = Node.objects.filter(Q(label__icontains=q) | Q(details__icontains=q))
+    results = Node.objects.filter(Q(label__icontains=q) | Q(details__icontains=q),
+                                  user=request.user)
     return dict(q=q,results=results)
 
 @login_required
 def add(request,node_id):
     if node_id:
         p = get_object_or_404(Node,id=node_id)
+        if p.user.id != request.user.id:
+            return HttpResponse("you are not the owner of this node")
         p.add_child(user=request.user,label=request.POST.get('label',''))
         return HttpResponseRedirect(p.get_absolute_url())
     else:
@@ -39,6 +44,8 @@ def add(request,node_id):
 @login_required
 def edit(request,node_id):
     node = get_object_or_404(Node,id=node_id)
+    if node.user.id != request.user.id:
+        return HttpResponse("you are not the owner of this node")
     node.label = request.POST.get('label','')
     node.details = request.POST.get('details','')
     if request.POST.get('target',None):
@@ -54,6 +61,8 @@ def api(request,node_id):
         d = loads(request.read())
         if node_id:
             p = get_object_or_404(Node,id=node_id)
+            if p.user.id != request.user.id:
+                return HttpResponse("you are not the owner of this node")
             p.add_child(user=request.user,label=d['label'])
         else:
             n = Node.add_root(user=request.user,label=d['label'])
@@ -61,6 +70,8 @@ def api(request,node_id):
     if request.method == "PUT":
         if node_id:
             n = get_object_or_404(Node,id=node_id)
+            if n.user.id != request.user.id:
+                return HttpResponse("you are not the owner of this node")
             d = loads(request.read())
             n.label = d['label']
             n.details = d['details']
@@ -69,11 +80,17 @@ def api(request,node_id):
     if request.method == "DELETE":
         if node_id:
             n = get_object_or_404(Node,id=node_id)
+            if n.user.id != request.user.id:
+                return HttpResponse("you are not the owner of this node")
+
             n.delete()
             return HttpResponse(dumps({}),mimetype="application/json")
 
     if node_id:
         n = get_object_or_404(Node,id=node_id)
+        if n.user.id != request.user.id:
+            return HttpResponse("you are not the owner of this node")
+
         return HttpResponse(dumps([n.as_dict() for n in n.get_children()]),
                             mimetype="application/json")
     else:
@@ -85,6 +102,8 @@ def api(request,node_id):
 def reorder(request,node_id):
     if request.method == "POST":
         p = get_object_or_404(Node,id=node_id)
+        if p.user.id != request.user.id:
+            return HttpResponse("you are not the owner of this node")
         keys = [k for k in request.GET.keys() if k.startswith("node_")]
         keys.sort(key=lambda x:int(x.split("_")[1]))
         children = [int(request.GET[k]) for k in keys if k.startswith('node_')]
